backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
akshare==1.11.60
TA-Lib==0.4.28
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
APScheduler==3.10.4
aiohttp==3.9.1